% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDat.R
\name{genDat}
\alias{genDat}
\title{Generate a simulated probability sample with a nested validation sample}
\usage{
genDat(seedNum, p11, p10, valType, n = NULL, alphaInt = NULL)
}
\arguments{
\item{seedNum}{Seed used for random generation}

\item{p11}{Sensitivity, P(Y*=1 | Y=1)}

\item{p10}{Value for P(Y*=1 | Y=0), otherwise known as (1-Specificity)}

\item{valType}{Type of validation sample (SRS or non-probability)}

\item{n}{Size of entire probability sample}

\item{alphaInt}{Intercept value relevant for size of a non-probability validation sample}
}
\value{
Data frame containing error-free outcomes, misclassified outcomes,
and indices for individuals belonging to the validation sample
}
\description{
Generate a simulated probability sample with a nested validation sample
}
