% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induceME.R
\name{induceME}
\alias{induceME}
\title{Induce misclassification in a binary outcome}
\usage{
induceME(yVec, p11, p10)
}
\arguments{
\item{yVec}{Vector of outcomes}

\item{p11}{Sensitivity, P(Y*=1 | Y=1)}

\item{p10}{Value for P(Y*=1 | Y=0), otherwise known as (1-Specificity)}
}
\value{
Vector of outcomes subject to misclassification
}
\description{
Induce misclassification in a binary outcome
}
