% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_est4.R
\name{run_est4}
\alias{run_est4}
\title{Run approach from Equations 3 and 4 that uses gold standard outcomes from validation data
and all silver standard outcomes}
\usage{
run_est4(df, inVal, varTrt, varGold, varSilver, varTrtMod, varSelectMod, opt)
}
\arguments{
\item{df}{Data frame containing exposure status, gold-standard outcomes, silver-standard outcomes,
and all relevant variables for analyses}

\item{inVal}{Vector that indicates whether an individual was included in the validation data (=1 for included)}

\item{varTrt}{Variable name for treatment/exposure}

\item{varGold}{Variable name for gold-standard outcome}

\item{varSilver}{Variable name for silver-standard outcome}

\item{varTrtMod}{Variable names for modeling treatment propensity model}

\item{varSelectMod}{Variable names for modeling validation sample selection model}

\item{opt}{Whether to use the optimal weight; TRUE will use the optimal weight
while FALSE will weight by proportion of validation sample size}
}
\value{
Data frame containing weight used, final estimate of ATE, final variance estimate of ATE,
estimate of ATE using validation data, variance estimate of ATE using validation data,
estimate of ATE using silver standard outcomes, and variance estimate of ATE using silver standard outcomes
}
\description{
Run approach from Equations 3 and 4 that uses gold standard outcomes from validation data
and all silver standard outcomes
}
